MyTicketApp/inventario/permissions.py
from rest_framework import permissions


class IsSuperUser(permissions.BasePermission):
    def has_permission(self, request, view):
        if request.user and request.user.groups.filter(name='SuperUser'):
            return True
        return False
